// src/models/banners.model.ts
export interface BANNERS  {
    banners_id?:  number;
    banners_title:  string;
    banners_url:  string;
    banners_image:  string;
    banners_group:  string;
    banners_html_text:  string;
    expires_impressions:  number;
    expires_date:  string;
    date_scheduled:  string;
    date_added:  string;
    date_status_change:  string;
    status:  number
}